refactor: move buildTokenBasedSettlement into settlementBuilder
