+++ src/apps/explorer/components/TransanctionBatchGraph/settlementBuilder.tsx
@@ -1,17 +1,21 @@
 import {
+  Account,
+  accountAddressesInvolved,
+  Contract,
+  getAliasFromAddress,
+  traceToTransfersAndTrades,
   Trade,
   Transfer,
-  Account,
-  getAliasFromAddress,
-  Contract,
-  traceToTransfersAndTrades,
-  accountAddressesInvolved,
 } from 'api/tenderly'
 import { SingleErc20State } from 'state/erc20'
 import BigNumber from 'bignumber.js'
 import { getExplorerUrl } from 'utils/getExplorerUrl'
 import { ContractTrade } from 'apps/explorer/components/TransanctionBatchGraph/alternativeView'
-import { BuildSettlementParams } from 'apps/explorer/components/TransanctionBatchGraph/alternativeView/hooks'
+import { TransactionData } from 'hooks/useTransactionData'
+import { Network } from 'types'
+import { Order } from 'api/operator'
+import { getContractTrades, getTokenAddress } from './alternativeView'
+import { abbreviateString } from 'utils'
 
 type Dict<T> = Record<string, T>
 
@@ -127,5 +131,62 @@ export function buildContractBasedSettlement(params: BuildSettlementParams): Set
     tokens: filteredTokens,
     trades,
     accounts,
+  }
+}
+
+export type BuildSettlementParams = {
+  networkId: Network | undefined
+  tokens: Dict<SingleErc20State>
+  orders?: Order[] | undefined
+  txData: TransactionData
+}
+
+export function buildTokenBasedSettlement(params: BuildSettlementParams): Settlement | undefined {
+  const { networkId, txData, tokens } = params
+  const { trace, contracts } = txData
+
+  if (!networkId || !trace || !contracts) {
+    return undefined
+  }
+
+  const { trades, transfers } = traceToTransfersAndTrades(trace)
+  const contractTrades = getContractTrades(trades, transfers)
+
+  const addressesSet = transfers.reduce((set, transfer) => {
+    set.add(getTokenAddress(transfer.token, networkId || 1))
+    return set
+  }, new Set<string>())
+
+  const tokenAddresses = Array.from(addressesSet)
+
+  const accounts = tokenAddresses.reduce((acc, address) => {
+    const symbol = tokens?.[address]?.symbol
+
+    acc[address] = {
+      alias: symbol || abbreviateString(address, 6, 4),
+      address,
+      href: getExplorerUrl(networkId, 'token', address),
+    }
+
+    return acc
+  }, {})
+
+  const filteredTokens = tokenAddresses.reduce((acc, address) => {
+    const token = tokens[address]
+
+    if (token) {
+      acc[address] = token
+    }
+
+    return acc
+  }, {})
+
+  return {
+    accounts,
+    trades,
+    contractTrades,
+    transfers,
+    tokens: filteredTokens,
+    contracts,
   }
 }
